pyproject.toml
[project]
name = "dnn-project"
version = "0.1.0"
description = "DINOv3 Vision Transformer Project"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "mlflow>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["model", "project_tracker"]
